Test Script Changs

+++ test/cache/CacheTable/test022.ksh
@@ -11,9 +11,16 @@ if [ -s "$input" ]
 then
     REL_PATH=${PWD}/cache/CacheTable
 fi
-export CSQL_CONFIG_FILE=$REL_PATH/csql.conf
-echo DSN=$DSN >> $CSQL_CONFIG_FILE
+rm -f /tmp/csql.conf
+cp $REL_PATH/csql.conf /tmp
+export CSQL_CONFIG_FILE=/tmp/csql.conf
+echo DSN=$DSN >>$CSQL_CONFIG_FILE
+
 isql $DSN < ${REL_PATH}/createt5.sql >/dev/null 2>&1
+if [ $? -ne 0 ]
+then
+    exit 1;
+fi
 echo table t5 is created with records in target db
 
 rm -f /tmp/csql/csqltable.conf /tmp/csql/csql.db
@@ -30,7 +37,7 @@ then
     exit 1;
 fi
 
-$CSQL_INSTALL_ROOT/bin/catalog -u root -p manager -D chunk
+$CSQL_INSTALL_ROOT/bin/catalog -u root -p manager -l
 if [ $? -ne 0 ]
 then
     exit 2;
